--- src/services/users/userServices.js
@@ -1,21 +1,0 @@
-import axios from "axios"
-
-export const getAllUser = async () => {
-  return await axios.get("http://localhost:5050/user/all").then(res => res).catch(err => err)
-}
-
-export const getOneUser = async (id) => {
-  return await axios.get(`http://localhost:5050/user/${id}`).then(res => res).catch(err => err)
-}
-
-export const updateOneUser = async (id, data) => {
-  return await axios.put(`http://localhost:5050/user/edit/${id}`, {...data}).then(res => res).catch(err => err)
-}
-
-export const deleteUser = async (id) => {
-  return await axios.delete(`http://localhost:5050/user/${id}`).then(res => res).catch(err => err)
-}
-
-export const addUser = async (data) => {
-  return await axios.post(`http://localhost:5050/user/add`, {...data})
-}
